+++ common/src/main/java/com/teno/openmarket/common/security/TokenBlacklistValidator.java
@@ -1,0 +1,5 @@
+package com.teno.openmarket.common.security;
+
+public interface TokenBlacklistValidator {
+    boolean existsByAccessToken(String token);
+}
