+++ user/src/main/java/com/teno/openmarket/user/feature/reissue/ReissueService.java
@@ -4,7 +4,7 @@ import com.teno.openmarket.common.error.GlobalErrorCode;
 import com.teno.openmarket.common.exception.BusinessException;
 import com.teno.openmarket.user.domain.token.RefreshToken;
 import com.teno.openmarket.user.domain.token.RefreshTokenRepository;
-import com.teno.openmarket.user.infra.security.JwtTokenProvider;
+import com.teno.openmarket.common.security.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
