+++ user/src/test/java/com/teno/openmarket/user/feature/logout/LogoutApiTest.java
@@ -2,7 +2,7 @@ package com.teno.openmarket.user.feature.logout;
 
 import com.teno.openmarket.test.support.BaseControllerTest;
 import com.teno.openmarket.user.config.TestUserSecurityConfig;
-import com.teno.openmarket.user.infra.security.JwtTokenProvider;
+import com.teno.openmarket.common.security.JwtTokenProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.context.annotation.Import;
