+++ user/src/test/java/com/teno/openmarket/user/feature/reissue/ReissueServiceTest.java
@@ -4,7 +4,7 @@ import com.teno.openmarket.common.error.GlobalErrorCode;
 import com.teno.openmarket.common.exception.BusinessException;
 import com.teno.openmarket.user.domain.token.RefreshToken;
 import com.teno.openmarket.user.domain.token.RefreshTokenRepository;
-import com.teno.openmarket.user.infra.security.JwtTokenProvider;
+import com.teno.openmarket.common.security.JwtTokenProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
